% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Celsius_Conversions.R
\name{C_to_F}
\alias{C_to_F}
\title{C_to_F Celsius Conversion}
\usage{
C_to_F(C_temp)
}
\arguments{
\item{F_temp}{The temperature in degrees Celsius}
}
\value{
The temperature in degrees Fahrenheit
}
\description{
This function allows you convert Celsius to Fahrenheit
}
\examples{

temp1 <-C_to_F(22);
temp2 <-C_to_F(c(-2,12,23));
}
