% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fahrenheit_Conversions.R
\name{F_to_C}
\alias{F_to_C}
\title{F_to_C Fahrenheit Conversion}
\usage{
F_to_C(F_temp)
}
\arguments{
\item{F_temp}{The temperature in degrees Fahrenheit}
}
\value{
The temperaure in degrees Celsius
}
\description{
This function allows you convert Fahrenheit to Celsius
}
\examples{

temp1 <-F_to_C(50);
temp2 <-F_to_C(c(50,60,30));
}
